requirements.txt
weaviate-client==4.5.4
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-text-splitters==0.0.1
pandas==2.0.3
sentence-transformers==2.5.1

